+++ extra/files/gen2-linux/ansible.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr//bin/env bash
 
 usage() { echo "Usage: $0 [-i <true|false> Install or uninstall ansible ]" 1>&2; }
 
@@ -33,17 +33,17 @@ echo "Found os: $OS"
 
 function install_ansible {
   /usr/bin/python3 -m pip install --upgrade pip
-  /usr/bin/python3 -m pip install --upgrade jmespath jsonlint yamllint ansible-base ansible pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp
-  /usr/local/bin/ansible-galaxy collection install ansible.posix
-  /usr/local/bin/ansible-galaxy collection install community.general
-  /usr/local/bin/ansible-galaxy collection install community.crypto
+  /usr/bin/python3 -m pip install --upgrade jmespath jsonlint yamllint ansible-core ansible pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp
+  /usr/local/bin/ansible-galaxy collection install --upgrade ansible.posix
+  /usr/local/bin/ansible-galaxy collection install --upgrade community.general
+  /usr/local/bin/ansible-galaxy collection install --upgrade community.crypto
 }
 
 if [ "$INSTALL" == "true" ] && [[ "$OS" =~ rhel|centos|fodora ]];then
   echo "Installing ansible on RHEL/related"
   yum clean all -y
   yum makecache -y
-  yum remove ansible ansible-base -y||true
+  yum remove ansible ansible-base ansible-core -y||true
   yum install python3 python3-devel python3-pip python3-wheel krb5-devel krb5-workstation -y
   yum install cowsay -y
   yum install python3-setuptools python3-psutil -y
@@ -55,7 +55,7 @@ if [ "$INSTALL" == "false" ] && [[ "$OS" =~ rhel|centos|fedora ]];then
   echo "Removing ansible on RHEL/related"
   yum clean all -y
   yum makecache -y
-  /usr/bin/python3 -m pip uninstall jmespath jsonlint yamllint ansible-base ansible setuptools-rust pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp -y
+  /usr/bin/python3 -m pip uninstall jmespath jsonlint yamllint ansible-core ansible setuptools-rust pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp -y
   rm -rfv /root/.ansible||true
   rm -rfv /root/.cache||true
   rm -rfv /home/vagrant/.ansible||true
@@ -67,7 +67,7 @@ if [ "$INSTALL" == "true" ] && [[ "$OS" =~ debian|ubuntu ]];then
   echo "Installing ansible on Ubuntu/Debian"
   apt-get clean all -y
   apt-get update -y
-  apt-get purge ansible ansible-base -y||true
+  apt-get purge ansible ansible-base ansible-core -y||true
   apt-get install python3 python3-dev python3-pip python3-wheel libkrb5-dev -y
   apt-get install cowsay -y
   install_ansible
@@ -77,11 +77,9 @@ if [ "$INSTALL" == "false" ] && [[ "$OS" =~ debian|ubuntu ]];then
   echo "Removing ansible on Ubuntu/Debian"
   apt-get clean all -y
   apt-get update -y
-  /usr/bin/python3 -m pip uninstall  jmespath jsonlint yamllint ansible-base ansible pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp -y
+  /usr/bin/python3 -m pip uninstall jmespath jsonlint yamllint ansible-core ansible pywinrm requests-kerberos requests-ntlm requests-credssp pypsrp -y
   rm -rfv /root/.ansible||true
   rm -rfv /root/.cache||true
   rm -rfv /home/vagrant/.ansible||true
   rm -rfv /home/vagrant/.cache||true
 fi
-
-