+++ extra/scripts/phase-4.windows-updates.ps1
@@ -2,41 +2,41 @@
 #install Windows Updates
 # For test run disable updates
 # exit
-Write-Output "Start Phase-4 - Updates"
+Write-Output "Phase-4 [START] - Updates"
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
 try {
-    Write-Output "Start Phase-4 - Installing Nuget"
+    Write-Output "Phase-4 [INFO] - Installing Nuget"
     Get-PackageProvider -Name "Nuget" -ForceBootstrap -Verbose -ErrorAction Stop
 }
 catch {
-    Write-Output "Start Phase-4 - Installation of nuget failed, exiting"
+    Write-Output "Phase-4 [WARN] - Installation of nuget failed, exiting"
 }
 # workaround for lastest PSWindowsUpdate
 
 try {
-    Write-Output "Start Phase-4 - Installing PSWindowsUpdate"
+    Write-Output "Phase-4 [INFO] - Installing PSWindowsUpdate"
     Install-Module PSWindowsUpdate -Force -Confirm:$false -Verbose -ErrorAction Stop
     Add-WUServiceManager -ServiceID 7971f918-a847-4430-9279-4a52d1efe18d -Confirm:$false
 }
 catch {
-    Write-Output "Start Phase-4 - Installation of PSWindowsUpdate failed, exiting"
+    Write-Output "Phase-4 [INFO]- Installation of PSWindowsUpdate failed, exiting"
     exit (1)
 }
 
 try {
     $modversion=((get-module -listavailable -name pswindowsupdate -ErrorAction SilentlyContinue).version.major|select-object -first 1)
     if ($modversion -gt 1) {
-        Write-Output "End Phase-4 - Updates pass started"
+        Write-Output "Phase-4 [INFO] - Updates pass started"
         Install-WindowsUpdate -AcceptAll -MicrosoftUpdate -IgnoreReboot -ErrorAction SilentlyContinue
-        Write-Output "End Phase-4 - Updates pass completed"
+        Write-Output "Phase-4 [INFO] - Updates pass completed"
         }
     else {
         Get-WUInstall -MicrosoftUpdate -AcceptAll -IgnoreReboot -ErrorAction SilentlyContinue
         }
 }
 catch {
-    Write-Output "End Phase-4 - Updates pass failed, not critical"
+    Write-Output "Phase-4 [WARN] - Updates pass failed, not critical"
     exit (0)
 }
-Write-Output "End Phase-4 - Updates"
+Write-Output "Phase-4 [END] - Updates"
 exit 0